+++ CHANGELOG.md
@@ -29,7 +29,10 @@ For older changes see the [archived Singularity change log](https://github.com/a
   the `--underlay` flag when overlay is implied for bind mounts but the
   kernel is too old to support fuse mounts in user namespaces and so
   tries to use fusermount.
-- Allow a writable `--overlay` to be used with `--nvccli` instead of `--writable-tmpfs`
+- When someone uses a `yum` bootstrap to build a container without using
+  subuid-based fakeroot or root, warn that it is unlikely to work.
+- Allow a writable `--overlay` to be used with `--nvccli` instead of
+  `--writable-tmpfs`.
 
 ## v1.3.3 - \[2024-07-03\]